refactor: rename ConsolidatedProgressionChanges to ConsolidatedReleasePlanChanges
